+++ common/src/main/java/traben/entity_texture_features/config/ETFConfig.java
@@ -50,6 +50,8 @@ public class ETFConfig {
 
     public Set<ETFConfigScreenWarnings.ConfigWarning> ignoredConfigs = new HashSet<>();
 
+    public boolean hideConfigButton = false;
+
     //string name stuff more in-depth than other enum for backwards compatibility
 
     public static ETFConfig copyFrom(ETFConfig source) {